split sym_init
